starterFiles/requirements.txt
absl-py==2.0.0
alembic==1.12.1
aniso8601==9.0.1
anyio==4.0.0
apispec==5.2.2
apispec-webframeworks==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Automat==22.10.0
Babel==2.13.1
beautifulsoup4==4.12.2
bidict==0.22.1
binaryornot==0.4.4
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.0
constantly==23.10.4
contourpy==1.2.0
cookiecutter==2.1.1
cycler==0.12.1
dask==2023.11.0
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.2.3
defusedxml==0.7.1
distributed==2023.11.0
dnspython==2.4.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.18.1
Flask==3.0.0
Flask-Cors==4.0.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.44.1
fqdn==1.5.1
fsspec==2023.10.0
gast==0.5.4
gevent==23.9.1
gevent-websocket==0.10.1
gitignore-parser==0.1.9
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.2
h11==0.14.0
h5py==3.10.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jinja2-time==0.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
keras==2.14.0
kiwisolver==1.4.5
kthread==0.2.3
libclang==16.0.6
locket==1.0.0
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.28.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
modin==0.25.0
msgpack==1.0.7
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==2.8.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
passlib==1.7.4
Pillow==10.1.0
platformdirs==4.0.0
prometheus-client==0.18.0
prompt-toolkit==3.0.41
protobuf==4.25.0
psutil==5.9.6
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pymongo==4.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-json-logger==2.0.7
python-slugify==8.0.1
python-socketio==5.10.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
rsa==4.9
Send2Trash==1.8.2
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
taipy==3.0.0
taipy-config==3.0.0
taipy-core==3.0.0
taipy-gui==3.0.0
taipy-rest==3.0.0
taipy-templates==3.0.0
tblib==3.0.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.0
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.13.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
zict==3.0.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
